remove lockups, keep lp model
