+++ insurance-fund/programs/insurance-fund/src/instructions/admin/initialize_insurance_fund.rs
@@ -1,15 +1,10 @@
 use anchor_lang::prelude::*;
 use crate::states::*;
 use crate::constants::*;
-use crate::errors::InsuranceFundError;
 use crate::events::InitializeInsuranceFund as InitializeInsuranceFundEvent;
 
 #[derive(AnchorSerialize, AnchorDeserialize)]
 pub struct InitializeInsuranceFundArgs {
-    pub cold_wallet: Pubkey,
-    pub hot_wallet_share_bps: u64,
-    pub cold_wallet_share_bps: u64,
-    pub reward_mint: Pubkey,
     pub cooldown_duration: u64,
 }
 
@@ -19,17 +14,8 @@ pub fn initialize_insurance_fund(
 ) -> Result<()> {
 
     let InitializeInsuranceFundArgs {
-        cold_wallet,
-        cold_wallet_share_bps,
-        hot_wallet_share_bps,
-        reward_mint,
         cooldown_duration
     } = args;
-
-    require!(
-        cold_wallet_share_bps + hot_wallet_share_bps == 10_000,
-        InsuranceFundError::ShareConfigOverflow
-    );
 
     let signer = &ctx.accounts.signer;
     let admin = &mut ctx.accounts.admin;
@@ -42,17 +28,10 @@ pub fn initialize_insurance_fund(
     settings.bump = ctx.bumps.settings;
     settings.lockups = 0;
     settings.cooldown_duration = cooldown_duration;
-    settings.cold_wallet = cold_wallet;
-    settings.shares_config = SharesConfig {
-        cold_wallet_share_bps,
-        hot_wallet_share_bps
-    };
     settings.reward_config = RewardConfig {
         main: reward_mint
     };
     settings.frozen = false;
-    settings.superadmins = 1;
-    settings.debt_records = 0;
 
     emit!(InitializeInsuranceFundEvent {
         new_admin: signer.key()
