+++ insurance-fund/programs/insurance-fund/src/instructions/admin/mod.rs
@@ -4,12 +4,6 @@ pub use add_asset::*;
 pub mod initialize_insurance_fund;
 pub use initialize_insurance_fund::*;
 
-pub mod initialize_lockup;
-pub use initialize_lockup::*;
-
-pub mod initialize_lockup_vaults;
-pub use initialize_lockup_vaults::*;
-
 pub mod boost_rewards;
 pub use boost_rewards::*;
 
@@ -19,14 +13,8 @@ pub use manage_freeze::*;
 pub mod add_admin;
 pub use add_admin::*;
 
-pub mod remove_admin;
-pub use remove_admin::*;
-
 pub mod update_deposit_cap;
 pub use update_deposit_cap::*;
-
-pub mod rebalance;
-pub use rebalance::*;
 
 pub mod initialize_lp;
 pub use initialize_lp::*;
