+++ insurance-fund/programs/insurance-fund/src/instructions/swap/mod.rs
@@ -1,11 +1,2 @@
 pub mod swap;
 pub use swap::*;
-
-pub mod borrow;
-pub use borrow::*;
-
-pub mod repay;
-pub use repay::*;
-
-pub mod swap_lp;
-pub use swap_lp::*;
