--- insurance-fund/programs/insurance-fund/src/states/debt_record.rs
@@ -1,10 +1,0 @@
-use anchor_lang::prelude::*;
-
-#[account]
-#[derive(InitSpace)]
-pub struct DebtRecord {
-    pub amount: u64,
-    pub asset: Pubkey,
-    pub lockup: u64,
-    pub timestamp: u64,
-}
