+++ insurance-fund/programs/insurance-fund/src/states/mod.rs
@@ -1,9 +1,6 @@
 pub mod settings;
 pub use settings::*;
 
-pub mod lockup;
-pub use lockup::*;
-
 pub mod deposit;
 pub use deposit::*;
 
@@ -21,9 +18,6 @@ pub use cooldown::*;
 
 pub mod admin;
 pub use admin::*;
-
-pub mod debt_record;
-pub use debt_record::*;
 
 pub mod lp;
 pub use lp::*;
